Bug fixed and Project Completed. Final Commit for now

+++ 8_16384/DirPage.h
@@ -3,7 +3,7 @@
 class DirPage{
     public:
         struct DirPageHeader header;
-        uint32_t data_offsets[offsets_per_dir];
+        uint64_t data_offsets[offsets_per_dir];
         DirPage();
         bool Insert(size_t offset);
         void print_page_info();
